Remove duplication in tests RepositoryFixProcessingActionTests
